refactor: split useQuery hooks into custom hooks ♻️
