+++ apps/core/src/routes/dashboard/_layout/$tableName/route.tsx
@@ -1,8 +1,7 @@
 import LoadingSpinner from "@/components/loading-spinner"
-import { useQuery } from "@tanstack/react-query"
+import { useGetTableColumns } from "@/hooks/table"
 import { createFileRoute } from "@tanstack/react-router"
 import { z } from "zod"
-import { generateColumnsDefs } from "./-components/columns"
 import DataTable from "./-components/data-table"
 
 const tablePageSchema = z.object({
@@ -18,12 +17,11 @@ export const Route = createFileRoute("/dashboard/_layout/$tableName")({
 function TableData() {
   const { tableName } = Route.useParams()
 
-  const { data: columns, isLoading: isColumnsLoading } = useQuery({
-    queryKey: ["table_columns", tableName],
-    queryFn: async () => await generateColumnsDefs(tableName)
-  })
+  const { data: columns, isLoading: isColumnsLoading } =
+    useGetTableColumns(tableName)
 
   if (isColumnsLoading) return <LoadingSpinner />
+
   return (
     <section className="flex w-full flex-col overflow-auto will-change-scroll">
       <DataTable columns={columns!} tableName={tableName} />
